Following modifications done in TrackPredictor. 1) In Coordinated we are NOT storing z of point while converting to string. 2) A function added to DetectorMapping, to get Z of Detector for specified type. 3) Changing int to long long int in Global.h, so that we can collect more data. 4) Added TestTrackPredictorWithSimulatedData.cpp so that we can fill all the    combinations from the simulated data.
Current the processing time of making Combination increases exponential as
we increase the number of events.

So the code need improvement. One place of improvement is, instead of doing
linear search in the combination vector, one should try binary search.
This will result in less search time.

+++ base/Global.h
@@ -69,8 +69,8 @@ namespace Tomography{
             return fInstance;
         }
     }
-    static int fNumOfEvents ;
-    static int fEffEvNo ;
+    static unsigned long long int fNumOfEvents ;
+    static unsigned long long int fEffEvNo ;
 
     static bool BreakSimulation(){
         return fEffEvNo > fNumOfEvents ;
